chore(meshes): removes nav tabs from detail views
Renames MeshOverview to MeshDetailView to follow the pattern we have established for all other components of this type.
