+++ src/app/meshes/views/MeshDetailView.vue
