+++ src/app/meshes/views/MeshIndexView.vue
@@ -1,17 +1,12 @@
 <template>
-  <RouteView
-    v-slot="{route}"
-  >
-    <RouteTitle
-      :title="t('meshes.routes.item.title', {name: route.params.mesh})"
-    />
+  <RouteView>
     <AppView
       :breadcrumbs="[
         {
           to: {
             name: 'mesh-list-view'
           },
-          text: t('meshes.routes.item.breadcrumbs')
+          text: t('meshes.routes.item.breadcrumbs'),
         },
         {
           to: {
@@ -20,13 +15,20 @@
               mesh: route.params.mesh,
             },
           },
-          text: route.params.mesh
+          text: route.params.mesh as string,
         }
       ]"
     >
-      <RouterView
-        v-slot="child"
-      >
+      <template #title>
+        <h1>
+          <RouteTitle
+            :title="t('meshes.routes.item.title', { name: route.params.mesh as string })"
+            :render="true"
+          />
+        </h1>
+      </template>
+
+      <RouterView v-slot="child">
         <component
           :is="child.Component"
           :key="child.route.path"
@@ -35,11 +37,25 @@
     </AppView>
   </RouteView>
 </template>
+
 <script lang="ts" setup>
+import { watch } from 'vue'
+import { useRoute } from 'vue-router'
+
 import AppView from '@/app/application/components/app-view/AppView.vue'
 import RouteTitle from '@/app/application/components/route-view/RouteTitle.vue'
 import RouteView from '@/app/application/components/route-view/RouteView.vue'
+import { useStore } from '@/store/store'
 import { useI18n } from '@/utilities'
-const { t } = useI18n()
 
+const { t } = useI18n()
+const route = useRoute()
+const store = useStore()
+
+// Updates the policy type totals based on the current mesh.
+watch(() => route.params.mesh, (newMesh, oldMesh) => {
+  if (newMesh !== oldMesh && newMesh) {
+    store.dispatch('fetchPolicyTypeTotals', newMesh)
+  }
+}, { immediate: true })
 </script>
