+++ src/app/meshes/views/MeshTabsView.vue
@@ -1,17 +1,12 @@
 <template>
   <RouteView>
     <AppView>
-      <template #title>
-        <h1>{{ route.params.mesh }}</h1>
-      </template>
       <NavTabs
         class="route-mesh-view-tabs"
         :tabs="tabs"
       />
 
-      <RouterView
-        v-slot="child"
-      >
+      <RouterView v-slot="child">
         <component
           :is="child.Component"
           :key="child.route.path"
@@ -22,21 +17,17 @@
 </template>
 
 <script lang="ts" setup>
-import { watch } from 'vue'
-import { useRoute, useRouter, RouteRecordRaw } from 'vue-router'
+import { useRouter, RouteRecordRaw } from 'vue-router'
 
 import AppView from '@/app/application/components/app-view/AppView.vue'
 import RouteView from '@/app/application/components/route-view/RouteView.vue'
 import NavTabs, { NavTab } from '@/app/common/NavTabs.vue'
-import { useStore } from '@/store/store'
 import { useI18n } from '@/utilities'
 
 const i18n = useI18n()
-const route = useRoute()
 const router = useRouter()
-const store = useStore()
 
-const meshRoutes = router.getRoutes().find((route) => route.name === 'mesh-detail-view')?.children ?? []
+const meshRoutes = router.getRoutes().find((route) => route.name === 'mesh-tabs-view')?.children ?? []
 const tabs: NavTab[] = meshRoutes.map((route) => {
   const referenceRoute = typeof route.name === 'undefined' ? route.children?.[0] as RouteRecordRaw : route
   const routeName = referenceRoute.name as string
@@ -45,11 +36,4 @@ const tabs: NavTab[] = meshRoutes.map((route) => {
 
   return { title, routeName, module }
 })
-
-// Updates the policy type totals based on the current mesh.
-watch(() => route.params.mesh, (newMesh, oldMesh) => {
-  if (newMesh !== oldMesh && newMesh) {
-    store.dispatch('fetchPolicyTypeTotals', newMesh)
-  }
-}, { immediate: true })
 </script>
